refactor: remove progress references from tests and delete progress types
