+++ pkg/commands/compute/serve_test.go
@@ -67,7 +67,6 @@ func TestGetViceroy(t *testing.T) {
 
 	var out bytes.Buffer
 
-	progress := text.NewQuietProgress(&out)
 	av := mock.AssetVersioner{
 		AssetVersion:   "1.2.3",
 		BinaryFilename: viceroyBinName,
@@ -92,18 +91,14 @@ func TestGetViceroy(t *testing.T) {
 		ErrLog: fsterr.MockLog{},
 	}
 
-	_, err = compute.GetViceroy(progress, &out, av, &g)
+	spinner, err := text.NewSpinner(&out)
 	if err != nil {
 		t.Fatal(err)
 	}
-
-	// NOTE: We have to call progress.Done() here to prevent a data race because
-	// the getViceroy() function itself doesn't call it and so if we try to read
-	// from the shared bytes.Buffer (as we do below to validate its content)
-	// before calling Done(), then we'll get a race condition (this only shows up
-	// when running the complete test suite or this specific test with a -count
-	// value of 20 or above).
-	progress.Done()
+	_, err = compute.GetViceroy(spinner, &out, av, &g)
+	if err != nil {
+		t.Fatal(err)
+	}
 
 	if !strings.Contains(out.String(), "Fetching latest Viceroy release") {
 		t.Fatalf("expected file to be downloaded successfully")
